+++ src/Testing/FakeConnection.php
@@ -2,41 +2,40 @@
 
 declare(strict_types=1);
 
-namespace Lacasera\ElasticBridge\Tests;
+namespace Lacasera\ElasticBridge\Testing;
 
+use Elastic\Elasticsearch\Client;
 use Elastic\Elasticsearch\ClientBuilder;
 use Elastic\Elasticsearch\Exception\AuthenticationException;
 use Elastic\Elasticsearch\Response\Elasticsearch;
-use Http\Mock\Client;
+use Http\Mock\Client as MockClient;
 use Lacasera\ElasticBridge\Connection\ConnectionInterface;
 use Nyholm\Psr7\Response;
 use Override;
 
-class MockElasticConnection implements ConnectionInterface
+class FakeConnection implements ConnectionInterface
 {
-    protected \Elastic\Elasticsearch\Client $connection;
+    protected Client $connection;
 
     /**
      * @throws AuthenticationException
      */
     public function __construct(array $response, int $status = 200)
     {
-        $client = new Client;
+        $client = new MockClient;
 
         $this->connection = ClientBuilder::create()
             ->setHttpClient($client)
             ->build();
 
-        $response = new Response($status, [
+        $client->addResponse(new Response($status, [
             Elasticsearch::HEADER_CHECK => Elasticsearch::PRODUCT_NAME,
             'Content-Type' => 'application/json',
-        ], json_encode($response));
-
-        $client->addResponse($response);
+        ], json_encode($response)));
     }
 
     #[Override]
-    public function getClient(): \Elastic\Elasticsearch\Client
+    public function getClient(): Client
     {
         return $this->connection;
     }